+++ test/integration/chcol2/chcol2_test.go
@@ -1,0 +1,260 @@
+package chcol2
+
+import (
+	"context"
+	"os"
+	"path/filepath"
+	"syscall"
+	"time"
+
+	"github.com/anoideaopen/common-component/loggerhlp"
+	pbfound "github.com/anoideaopen/foundation/proto"
+	"github.com/anoideaopen/foundation/test/integration/cmn"
+	"github.com/anoideaopen/foundation/test/integration/cmn/client"
+	"github.com/anoideaopen/foundation/test/integration/cmn/fabricnetwork"
+	"github.com/anoideaopen/foundation/test/integration/cmn/runner"
+	"github.com/anoideaopen/glog"
+	"github.com/anoideaopen/robot/hlf"
+	"github.com/anoideaopen/robot/test/unit/common"
+	docker "github.com/fsouza/go-dockerclient"
+	"github.com/hyperledger/fabric/integration/nwo"
+	"github.com/hyperledger/fabric/integration/nwo/fabricconfig"
+	runnerFbk "github.com/hyperledger/fabric/integration/nwo/runner"
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	"github.com/onsi/gomega/gbytes"
+	"github.com/tedsuo/ifrit"
+	ginkgomon "github.com/tedsuo/ifrit/ginkgomon_v2"
+)
+
+const (
+	userName                = "backend"
+	channelWithoutChaincode = "empty"
+)
+
+var _ = Describe("Robot channel collector 2 tests", func() {
+	var (
+		testDir          string
+		cli              *docker.Client
+		network          *nwo.Network
+		networkProcess   ifrit.Process
+		ordererProcesses []ifrit.Process
+		peerProcesses    ifrit.Process
+	)
+
+	BeforeEach(func() {
+		networkProcess = nil
+		ordererProcesses = nil
+		peerProcesses = nil
+		var err error
+		testDir, err = os.MkdirTemp("", "foundation")
+		Expect(err).NotTo(HaveOccurred())
+
+		cli, err = docker.NewClientFromEnv()
+		Expect(err).NotTo(HaveOccurred())
+	})
+
+	AfterEach(func() {
+		if networkProcess != nil {
+			networkProcess.Signal(syscall.SIGTERM)
+			Eventually(networkProcess.Wait(), network.EventuallyTimeout).Should(Receive())
+		}
+		if peerProcesses != nil {
+			peerProcesses.Signal(syscall.SIGTERM)
+			Eventually(peerProcesses.Wait(), network.EventuallyTimeout).Should(Receive())
+		}
+		if network != nil {
+			network.Cleanup()
+		}
+		for _, ordererInstance := range ordererProcesses {
+			ordererInstance.Signal(syscall.SIGTERM)
+			Eventually(ordererInstance.Wait(), network.EventuallyTimeout).Should(Receive())
+		}
+		err := os.RemoveAll(testDir)
+		Expect(err).NotTo(HaveOccurred())
+	})
+
+	var (
+		channels         = []string{cmn.ChannelAcl, cmn.ChannelFiat, channelWithoutChaincode}
+		ordererRunners   []*ginkgomon.Runner
+		redisProcess     ifrit.Process
+		redisDB          *runner.RedisDB
+		networkFound     *cmn.NetworkFoundation
+		robotProc        ifrit.Process
+		skiBackend       string
+		peer             *nwo.Peer
+		admin            *client.UserFoundation
+		feeSetter        *client.UserFoundation
+		feeAddressSetter *client.UserFoundation
+	)
+
+	BeforeEach(func() {
+		By("start redis")
+		redisDB = &runner.RedisDB{}
+		redisProcess = ifrit.Invoke(redisDB)
+		Eventually(redisProcess.Ready(), runnerFbk.DefaultStartTimeout).Should(BeClosed())
+		Consistently(redisProcess.Wait()).ShouldNot(Receive())
+	})
+	BeforeEach(func() {
+		networkConfig := nwo.MultiNodeSmartBFT()
+		networkConfig.Channels = nil
+
+		pchs := make([]*nwo.PeerChannel, 0, cap(channels))
+		for _, ch := range channels {
+			pchs = append(pchs, &nwo.PeerChannel{
+				Name:   ch,
+				Anchor: true,
+			})
+		}
+		for _, peer := range networkConfig.Peers {
+			peer.Channels = pchs
+		}
+
+		network = nwo.New(networkConfig, testDir, cli, StartPort(), components)
+		cwd, err := os.Getwd()
+		Expect(err).NotTo(HaveOccurred())
+		network.ExternalBuilders = append(network.ExternalBuilders,
+			fabricconfig.ExternalBuilder{
+				Path:                 filepath.Join(cwd, ".", "externalbuilders", "binary"),
+				Name:                 "binary",
+				PropagateEnvironment: []string{"GOPROXY"},
+			},
+		)
+
+		networkFound = cmn.New(network, channels)
+		networkFound.Robot.RedisAddresses = []string{redisDB.Address()}
+		networkFound.ChannelTransfer.RedisAddresses = []string{redisDB.Address()}
+
+		networkFound.GenerateConfigTree()
+		networkFound.Bootstrap()
+
+		for _, orderer := range network.Orderers {
+			runner := network.OrdererRunner(orderer)
+			runner.Command.Env = append(runner.Command.Env, "FABRIC_LOGGING_SPEC=orderer.consensus.smartbft=debug:grpc=debug")
+			ordererRunners = append(ordererRunners, runner)
+			proc := ifrit.Invoke(runner)
+			ordererProcesses = append(ordererProcesses, proc)
+			Eventually(proc.Ready(), network.EventuallyTimeout).Should(BeClosed())
+		}
+
+		peerGroupRunner, _ := fabricnetwork.PeerGroupRunners(network)
+		peerProcesses = ifrit.Invoke(peerGroupRunner)
+		Eventually(peerProcesses.Ready(), network.EventuallyTimeout).Should(BeClosed())
+
+		By("Joining orderers to channels")
+		for _, channel := range channels {
+			fabricnetwork.JoinChannel(network, channel)
+		}
+
+		By("Waiting for followers to see the leader")
+		Eventually(ordererRunners[1].Err(), network.EventuallyTimeout, time.Second).Should(gbytes.Say("Message from 1"))
+		Eventually(ordererRunners[2].Err(), network.EventuallyTimeout, time.Second).Should(gbytes.Say("Message from 1"))
+		Eventually(ordererRunners[3].Err(), network.EventuallyTimeout, time.Second).Should(gbytes.Say("Message from 1"))
+
+		By("Joining peers to channels")
+		for _, channel := range channels {
+			network.JoinChannel(channel, network.Orderers[0], network.PeersWithChannel(channel)...)
+		}
+
+		peer = network.Peer("Org1", "peer0")
+
+		pathToPrivateKeyBackend := network.PeerUserKey(peer, "User1")
+		skiBackend, err = cmn.ReadSKI(pathToPrivateKeyBackend)
+		Expect(err).NotTo(HaveOccurred())
+
+		admin, err = client.NewUserFoundation(pbfound.KeyType_ed25519)
+		Expect(err).NotTo(HaveOccurred())
+		Expect(admin.PrivateKeyBytes).NotTo(Equal(nil))
+
+		feeSetter, err = client.NewUserFoundation(pbfound.KeyType_ed25519)
+		Expect(err).NotTo(HaveOccurred())
+		Expect(feeSetter.PrivateKeyBytes).NotTo(Equal(nil))
+
+		feeAddressSetter, err = client.NewUserFoundation(pbfound.KeyType_ed25519)
+		Expect(err).NotTo(HaveOccurred())
+		Expect(feeAddressSetter.PrivateKeyBytes).NotTo(Equal(nil))
+
+		cmn.DeployACL(network, components, peer, testDir, skiBackend, admin.PublicKeyBase58, admin.KeyType)
+		cmn.DeployFiat(network, components, peer, testDir, skiBackend,
+			admin.AddressBase58Check, feeSetter.AddressBase58Check, feeAddressSetter.AddressBase58Check)
+	})
+	BeforeEach(func() {
+		By("start robot")
+		robotRunner := networkFound.RobotRunner()
+		robotProc = ifrit.Invoke(robotRunner)
+		Eventually(robotProc.Ready(), network.EventuallyTimeout).Should(BeClosed())
+	})
+	AfterEach(func() {
+		By("stop robot")
+		if robotProc != nil {
+			robotProc.Signal(syscall.SIGTERM)
+			Eventually(robotProc.Wait(), network.EventuallyTimeout).Should(Receive())
+		}
+	})
+
+	It("Get data test", func() {
+		By("Test 3 new logger & context")
+		log, err := loggerhlp.CreateLogger("std", "debug")
+		Expect(err).NotTo(HaveOccurred())
+		logCtx := glog.NewContext(context.Background(), log)
+
+		By("Test 3 new channel collector creator")
+		chCr := hlf.NewChCollectorCreator(
+			cmn.ChannelFiat,
+			networkFound.ConnectionPath("User1"),
+			userName,
+			"Org1",
+			common.DefaultPrefixes,
+			1,
+		)
+		Expect(chCr).NotTo(BeNil())
+
+		dataReady := make(chan struct{}, 1)
+
+		By("Test 3 new channel collector")
+		chColl, err := chCr(logCtx, dataReady, cmn.ChannelFiat, 0)
+		Expect(err).NotTo(HaveOccurred())
+		Expect(chColl).NotTo(BeNil())
+
+		const closeAfterNBlocks = 1
+
+		By("Test 3 getting data")
+		for i := 0; ; i++ {
+			select {
+			case data, ok := <-chColl.GetData():
+				if ok {
+					log.Info("data from block:", data.BlockNum)
+				}
+
+				if !ok && i >= closeAfterNBlocks {
+					log.Info("channel closed i:", i)
+					return
+				}
+
+				Expect(ok).To(BeTrue())
+				Expect(data).NotTo(BeNil())
+
+				if !data.IsEmpty() {
+					log.Info("data for block:", data.BlockNum,
+						" lens:", len(data.Txs), len(data.Swaps), len(data.MultiSwaps), len(data.SwapsKeys), len(data.MultiSwapsKeys))
+				}
+
+			default:
+				log.Debug("need to wait")
+				<-dataReady
+			}
+
+			// after closeAfterNBlocks - init stop
+			if i == closeAfterNBlocks {
+				log.Info("schedule close")
+
+				// don't wait
+				go func() {
+					chColl.Close()
+					close(dataReady)
+					log.Info("was closed")
+				}()
+			}
+		}
+	})
+})
